pyproject.toml
[project]
name = "comfyui-polza"
description = "A ComfyUI nodes package providing access to 300+ AI models through a unified Polza.ai API"
version = "1.0.6"
license = { file = "LICENSE.txt" }
dependencies = [
    "requests>=2.28.0",
    "Pillow>=9.0.0",
    "numpy",
    "torch",
]

[project.urls]
Repository = "https://github.com/PioneerMNDR/ComfyUI-Polza"

[tool.comfy]
PublisherId = "pioneermndr"  # ID с сайта Registry https://comfyregistry.com
DisplayName = "Polza AI Nodes"
Icon = ""  # SVG, PNG, JPG или GIF — ссылка на иконку
